Reading single and all posts from the database.
The getPosts method that retrieve all existent posts in the database was
refactored to read the data from the database. Previously, the data
sent to the frontend was hardcoded.
There is a new route to read a single post based on an id pases as a
request parameter. The controller of this route query the data from the
database based on the id pased as a parameter.